♻️ Move `#build_metadata` to existing decorator
I forgot we already have a decorator for this. I'm moving the method
into it instead of making a new one.
